Feature: allow finding registered metric by metric name, type, or help (#86)
When reusing the same client in different places to export the same
metric, there is no external interface to retrieve this metric after
it's registered. Using `instance_variable_get` to reach into the
instance variable `@metrics` is less preferable. Re-registering the
metric will lead to an ever-growing array of `@metrics`.

This commit provides a way to get the metric by name, and optionally by
type and help, so that we can use the registered metric(s) to observe