Add support for HTTPAuth::BasicAuth in prometheus_exporter
Adds support for basic authentication in the prometheus_exporter CLI application.

When given a `htpasswd` filename via the `--auth` command line argument, the server will use this file to authenticate any requests to the `/metrics` endpoint.

Note: the htpasswd file must be created beforehand and due to limitations in WEBrick, only crypt is supported.

Optionally, a custom REALM may be provided via the `---realm` argument. The default REALM is "Prometheus Exporter"