Add back username and port as AR configuration values
In Rail 6.1 the information return is from an `ActiveRecord::DatabaseConfigurations::HashConfig` object, with methods for some but not all of the configuration attributes. This fixes a bug wheree `database` and `host` had getter methods, but the `username` and `port` values did not. All values are now just fetched from the instance variable configuration_hash directly.

Signed-off-by: Tim Chambers <tim@possibilogy.com>