Do not include status label everywhere
The `status` label is included in every metrics, and not only in
`http_requests_total`, even though the README states that
```
All metrics have a `controller` and an `action` label.
`http_requests_total` additionally has a (HTTP response) `status` label.
```

This can be a problem as this `status` label can quickly increase the
cardinality of the metrics.