fix bug, handle when queue is empty
When we do `group(:queue_name).size` it returns the count by queue
=> {"queue_a"=>3, "queue_d"=>1}

The problem is when a queue is empty it will simply be excluded from the
results instead of returning count of 0. So the result we want should be
=> {"queue_a"=>3, "queue_b"=>0, "queue_c"=>0, "queue_d"=>1}

Without returning 0, the queue count in prometheus metrics will be the
last non-zero value meaning we can't auto-scale down the workers.

Bug Fix Refactor