Require the main prometheus_exporter module in client and server modules
The main module includes some consts and defaults, which both client and server require to work.
In the README, it is suggested to treat client and server modules as entry-points,
but without requiring the main module they won't work.

In this commit, I'm adding a require and tests to show the problem.